Accept a reservation through an assignment callback
I am testing out a couple of workflows to see what will work best, in this
current iteration, I am responding to the assignmentCallback with an
instruction to dial the Agent. This initiates a custom callback which I have
also created /callbacks/ctc-agent-answers.

This is where we would say the TwiML and then dial-in the customer, assuming
we have all the necessary information from the task stored with it.