Customer creates a task instead of a phone call
For the purposes of task routing, when the customer submits a form, that form
should respond by creating a task with the necessary attributes and assigned
to the proper WorkflowSid. I've added the WorkflowSid to the necessary
configuration.

Since we can associate arbitrary attributes with Tasks, it feels unnecessary
to store the Call request for record keeping purposes. We can get most of this
information from the various status callbacks.